init/#1: rename package
