--- app/src/main/java/com/Kiero/KieroApplication.kt
@@ -1,26 +1,0 @@
-package com.Kiero
-
-import android.app.Application
-import androidx.appcompat.app.AppCompatDelegate
-import dagger.hilt.android.HiltAndroidApp
-import timber.log.Timber
-
-@HiltAndroidApp
-class KieroApplication : Application() {
-    override fun onCreate() {
-        super.onCreate()
-        setTimber()
-        setDayMode()
-    }
-
-    private fun setTimber() {
-        if (BuildConfig.DEBUG) {
-            Timber.Forest.plant(Timber.DebugTree())
-        }
-    }
-
-    private fun setDayMode() {
-        AppCompatDelegate.setDefaultNightMode(AppCompatDelegate.MODE_NIGHT_NO)
-    }
-
-}
