--- app/src/main/java/com/Kiero/core/model/UiState.kt
@@ -1,15 +1,0 @@
-package com.Kiero.core.model
-
-sealed interface UiState<out T> {
-    data object Empty : UiState<Nothing>
-
-    data object Loading : UiState<Nothing>
-
-    data class Success<T>(
-        val data: T,
-    ) : UiState<T>
-
-    data class Failure(
-        val message: String,
-    ) : UiState<Nothing>
-}
