--- app/src/main/java/com/Kiero/core/navigation/Route.kt
@@ -1,3 +1,0 @@
-package com.Kiero.core.navigation
-
-interface Route
