--- app/src/main/java/com/Kiero/core/network/di/NetworkModule.kt
@@ -1,70 +1,0 @@
-package com.Kiero.core.network.di
-
-import com.Kiero.BuildConfig
-import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
-import dagger.Module
-import dagger.Provides
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import kotlinx.serialization.json.Json
-import okhttp3.MediaType.Companion.toMediaType
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Converter
-import retrofit2.Retrofit
-import java.util.concurrent.TimeUnit
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-object NetworkModule {
-
-    @Provides
-    @Singleton
-    fun providesLoggingInterceptor() = HttpLoggingInterceptor().apply {
-        level = if (BuildConfig.DEBUG) {
-            HttpLoggingInterceptor.Level.BODY
-        } else {
-            HttpLoggingInterceptor.Level.NONE
-        }
-    }
-
-    @Provides
-    @Singleton
-    fun providesOkHttpClient(
-        loggingInterceptor: HttpLoggingInterceptor,
-    ): OkHttpClient = OkHttpClient.Builder()
-        .connectTimeout(10, TimeUnit.SECONDS)
-        .readTimeout(10, TimeUnit.SECONDS)
-        .addInterceptor(loggingInterceptor)
-        .build()
-
-
-    /*
-    Test 용 Setting 입니다
-     */
-    @Provides
-    @Singleton
-    fun providesJson(): Json = Json {
-        ignoreUnknownKeys = true
-        coerceInputValues = true
-        isLenient = true
-        prettyPrint = false
-        encodeDefaults = true
-    }
-
-    @Provides
-    @Singleton
-    fun providesConverterFactory(json: Json): Converter.Factory =
-        json.asConverterFactory("application/json".toMediaType())
-    @Provides
-    @Singleton
-    fun providesRetrofit(
-        client: OkHttpClient,
-        converterFactory: Converter.Factory,
-    ): Retrofit = Retrofit.Builder()
-        .baseUrl(BuildConfig.BASE_URL)
-        .addConverterFactory(converterFactory)
-        .client(client)
-        .build()
-}
