--- app/src/main/java/com/Kiero/data/auth/model/DummyEntity.kt
@@ -1,8 +1,0 @@
-package com.Kiero.data.auth.model
-
-data class DummyEntity(
-    val profile: String,
-    val firstName: String,
-    val id: Int,
-    val lastName: String,
-)
