--- app/src/main/java/com/Kiero/data/auth/remote/datasourceimpl/DummyDataSourceImpl.kt
@@ -1,14 +1,0 @@
-package com.Kiero.data.auth.remote.datasourceimpl
-
-import com.Kiero.core.network.model.DummyBaseResponse
-import com.Kiero.data.auth.remote.api.DummyService
-import com.Kiero.data.auth.remote.datasource.DummyDataSource
-import com.Kiero.data.auth.remote.dto.response.DummyResponseDto
-import javax.inject.Inject
-
-class DummyDataSourceImpl @Inject constructor(
-    private val dummyService: DummyService,
-) : DummyDataSource {
-    override suspend fun getDummyList(): DummyBaseResponse<List<DummyResponseDto>> =
-        dummyService.getDummyLists()
-}
