--- app/src/main/java/com/Kiero/data/auth/remote/dto/request/DummyRequestDto.kt
@@ -1,4 +1,0 @@
-package com.Kiero.data.auth.remote.dto.request
-
-class DummyRequestDto {
-}
