--- app/src/main/java/com/Kiero/data/auth/remote/dto/response/DummyResponseDto.kt
@@ -1,18 +1,0 @@
-package com.Kiero.data.auth.remote.dto.response
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class DummyResponseDto(
-    @SerialName("id")
-    val id: Int,
-    @SerialName("email")
-    val email: String,
-    @SerialName("first_name")
-    val firstName: String,
-    @SerialName("last_name")
-    val lastName: String,
-    @SerialName("avatar")
-    val avatar: String
-)
