--- app/src/main/java/com/Kiero/data/auth/repository/DummyRepository.kt
@@ -1,7 +1,0 @@
-package com.Kiero.data.auth.repository
-
-import com.Kiero.data.auth.model.DummyEntity
-
-interface DummyRepository {
-    suspend fun getDummyList(): Result<List<DummyEntity>>
-}
