--- app/src/main/java/com/Kiero/data/auth/repositoryimpl/DummyRepositoryImpl.kt
@@ -1,19 +1,0 @@
-package com.Kiero.data.auth.repositoryimpl
-
-import com.Kiero.core.common.util.suspendRunCatching
-import com.Kiero.data.auth.mapper.DummyMapper
-import com.Kiero.data.auth.remote.datasource.DummyDataSource
-import com.Kiero.data.auth.model.DummyEntity
-import com.Kiero.data.auth.repository.DummyRepository
-import javax.inject.Inject
-
-
-class DummyRepositoryImpl @Inject constructor(
-    private val dummyDataSource: DummyDataSource,
-    private val mapper: DummyMapper,
-) : DummyRepository {
-    override suspend fun getDummyList(): Result<List<DummyEntity>> = suspendRunCatching {
-        dummyDataSource.getDummyList().data.map { mapper.mapDtoToEntity(it) }
-    }
-}
-
