--- app/src/main/java/com/Kiero/data/di/RepositoryModule.kt
@@ -1,21 +1,0 @@
-package com.Kiero.data.di
-
-import com.Kiero.data.auth.repositoryimpl.DummyRepositoryImpl
-import com.Kiero.data.auth.repository.DummyRepository
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.components.SingletonComponent
-import javax.inject.Singleton
-
-@Module
-@InstallIn(SingletonComponent::class)
-interface RepositoryModule {
-
-    @Binds
-    @Singleton
-    fun bindsDummyRepository(
-        dummyRepositoryImpl: DummyRepositoryImpl
-    ): DummyRepository
-
-}
