+++ app/src/main/java/com/kiero/core/navigation/MainTabRoute.kt
@@ -1,0 +1,3 @@
+package com.kiero.core.navigation
+
+interface MainTabRoute : Route
