+++ app/src/main/java/com/kiero/core/network/model/DummyBaseResponse.kt
@@ -1,0 +1,18 @@
+package com.kiero.core.network.model
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class DummyBaseResponse<T>(
+    @SerialName("data")
+    val data: T,
+    @SerialName("page")
+    val page: Int? = null,
+    @SerialName("per_page")
+    val perPage: Int? = null,
+    @SerialName("total")
+    val total: Int? = null,
+    @SerialName("total_pages")
+    val totalPages: Int? = null
+)
