+++ app/src/main/java/com/kiero/data/auth/model/DummyEntity.kt
@@ -1,0 +1,8 @@
+package com.kiero.data.auth.model
+
+data class DummyEntity(
+    val profile: String,
+    val firstName: String,
+    val id: Int,
+    val lastName: String,
+)
