+++ app/src/main/java/com/kiero/data/auth/remote/dto/request/DummyRequestDto.kt
@@ -1,0 +1,4 @@
+package com.kiero.data.auth.remote.dto.request
+
+class DummyRequestDto {
+}
