+++ app/src/main/java/com/kiero/data/auth/repository/DummyRepository.kt
@@ -1,0 +1,7 @@
+package com.kiero.data.auth.repository
+
+import com.kiero.data.auth.model.DummyEntity
+
+interface DummyRepository {
+    suspend fun getDummyList(): Result<List<DummyEntity>>
+}
