+++ app/src/main/java/com/kiero/data/di/RepositoryModule.kt
@@ -1,0 +1,21 @@
+package com.kiero.data.di
+
+import com.kiero.data.auth.repositoryimpl.DummyRepositoryImpl
+import com.kiero.data.auth.repository.DummyRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+interface RepositoryModule {
+
+    @Binds
+    @Singleton
+    fun bindsDummyRepository(
+        dummyRepositoryImpl: DummyRepositoryImpl
+    ): DummyRepository
+
+}
