+++ app/src/main/java/com/kiero/data/di/ServiceModule.kt
@@ -1,0 +1,20 @@
+package com.kiero.data.di
+
+import com.kiero.data.auth.remote.api.DummyService
+import dagger.Module
+import dagger.Provides
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import retrofit2.Retrofit
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+object ServiceModule {
+
+    @Provides
+    @Singleton
+    fun providesDummyService(retrofit: Retrofit): DummyService =
+        retrofit.create(DummyService::class.java)
+
+}
